Various changes

+++ package_list_update.sh
@@ -25,6 +25,7 @@ fdp_errata_list_file=$script_directory/errata_list.txt
 package_list_file=~/package_list.txt
 upload_package_file=${upload_package_file:-"yes"}
 errata_list=${errata_list:-""}
+update_fdp_package_list=${update_fdp_package_list:-"no"}
 
 if [[ -z "$errata_list" ]]; then
 	rm -f ./batches.txt && touch ./batches.txt
@@ -40,9 +41,11 @@ sedeasy ()
 sed -i "s/$(echo $1 | sed -e 's/\([[\/.*]\|\]\)/\\&/g')/$(echo $2 | sed -e 's/[\/&]/\\&/g')/g" $3
 }
 
-rm -f $new_package_list_temp_file $new_package_list_file
-echo "" >> $new_package_list_file
-echo "# OVS-$fdp_release Packages" >> $new_package_list_file
+if [[ "$package_type" == "OVS" ]]; then
+	rm -f $new_package_list_temp_file $new_package_list_file
+	echo "" >> $new_package_list_file
+	echo "# $fdp_release Packages" >> $new_package_list_file
+fi
 
 pushd $script_directory
 
@@ -84,40 +87,57 @@ for i in $errata_list; do
 		package_url=$(grep packages $package_list_file | egrep -v '\-devel|ipsec|python|debug|test|scripts')
 		python_package_url=$(grep packages $package_list_file | grep python | egrep -v 'debug')
 		tcpdump_package_url=$(grep packages $package_list_file | grep 'noarch')
-		position_b=$(basename $package_url | awk -F - '{print $2}' | awk -F '.' '{print $1$2}')
-		position_b=$(echo $position_b"0")
-		url_rhel_ver=$(basename $package_url | awk -F 'el' '{print $2}' | awk -F 'fdp' '{print $1}')
-		echo "$package_type""$position_b"_"$fdp_release_short"_RHEL"$url_rhel_ver"=${package_url} >> $new_package_list_file
-		echo "$package_type""$position_b"_PYTHON_"$fdp_release_short"_RHEL"$url_rhel_ver"=${python_package_url} >> $new_package_list_file
-		echo "$package_type""$position_b"_TCPDUMP_"$fdp_release_short"_RHEL"$url_rhel_ver"=${tcpdump_package_url} >> $new_package_list_file
+		echo "OVS package URL: $package_url"
+		echo "OVS Python package URL: $python_package_url"
+		echo "OVS tcpdump package URL: $tcpdump_package_url"		
+		
+		# Steps below used for populating fdp_package_list.sh which may not be necessary
+		if [[ $update_fdp_package_list == "yes" ]]; then
+			position_b=$(basename $package_url | awk -F - '{print $2}' | awk -F '.' '{print $1$2}')
+			position_b=$(echo $position_b"0")
+			url_rhel_ver=$(basename $package_url | awk -F 'el' '{print $2}' | awk -F 'fdp' '{print $1}')
+			echo "$package_type""$position_b"_"$fdp_release_short"_RHEL"$url_rhel_ver"=${package_url} >> $new_package_list_file
+			echo "$package_type""$position_b"_PYTHON_"$fdp_release_short"_RHEL"$url_rhel_ver"=${python_package_url} >> $new_package_list_file
+			echo "$package_type""$position_b"_TCPDUMP_"$fdp_release_short"_RHEL"$url_rhel_ver"=${tcpdump_package_url} >> $new_package_list_file
+		fi
 	elif [[ "$package_type" == "OVN" ]]; then
 		ovn_common_package_url=$(grep packages $package_list_file | egrep -v '\-devel|central|host|vtep|debug')
 		ovn_central_package_url=$(grep packages $package_list_file | grep central | egrep -v '\-devel|debug')
 		ovn_host_package_url=$(grep packages $package_list_file | grep host | egrep -v '\-devel|debug')
-		position_b=$(basename $ovn_common_package_url | awk -F - '{print $2}' | awk -F '.' '{print $1$2}')
-		url_rhel_ver=$(basename $ovn_common_package_url | awk -F 'el' '{print $2}' | awk -F 'fdp' '{print $1}')
-		echo "$package_type"_COMMON_"$position_b"_"$fdp_release_short"_RHEL"$url_rhel_ver"=${ovn_common_package_url} >> $new_package_list_file
-		echo "$package_type"_CENTRAL_"$position_b"_"$fdp_release_short"_RHEL"$url_rhel_ver"=${ovn_central_package_url} >> $new_package_list_file
-		echo "$package_type"_HOST_"$position_b"_"$fdp_release_short"_RHEL"$url_rhel_ver"=${ovn_host_package_url} >> $new_package_list_file		
+		echo "OVS common package URL: $ovn_common_package_url"
+		echo "OVS central package URL: $ovn_central_package_url"
+		echo "OVS host package URL: $ovn_host_package_url"
+		
+		# Steps below used for populating fdp_package_list.sh which may not be necessary
+		if [[ $update_fdp_package_list == "yes" ]]; then
+			position_b=$(basename $ovn_common_package_url | awk -F - '{print $2}' | awk -F '.' '{print $1$2}')
+			url_rhel_ver=$(basename $ovn_common_package_url | awk -F 'el' '{print $2}' | awk -F 'fdp' '{print $1}')
+			echo "$package_type"_COMMON_"$position_b"_"$fdp_release_short"_RHEL"$url_rhel_ver"=${ovn_common_package_url} >> $new_package_list_file
+			echo "$package_type"_CENTRAL_"$position_b"_"$fdp_release_short"_RHEL"$url_rhel_ver"=${ovn_central_package_url} >> $new_package_list_file
+			echo "$package_type"_HOST_"$position_b"_"$fdp_release_short"_RHEL"$url_rhel_ver"=${ovn_host_package_url} >> $new_package_list_file
+		fi		
 	fi
 done
 
-# Remove any existing lines in $package_file referring to $fdp_release to avoid redundancy, etc
-existing_line_number=$(grep -n "$fdp_release" $package_file | awk -F ':' '{print $1}' | head -1)
-if [[ $existing_line_number ]]; then
-	echo "Removing existing entries for $fdp_release..."
-	line_number_before=$(( existing_line_number - 1 ))
-	line=$(sed -n "${line_number_before}p" $package_file)
+# Steps below used for modifying fdp_package_list.sh which may not be necessary
+if [[ $update_fdp_package_list == "yes" ]]; then
+	# Remove any existing lines in $package_file referring to $fdp_release to avoid redundancy, etc
+	existing_line_number=$(grep -n "$fdp_release" $package_file | awk -F ':' '{print $1}' | head -1)
+	if [[ $existing_line_number ]]; then
+		echo "Removing existing entries for $fdp_release..."
+		line_number_before=$(( existing_line_number - 1 ))
+		line=$(sed -n "${line_number_before}p" $package_file)
 
-	if [[ -z "$line" ]]; then
-		sed -n -i "1,$(( line_number_before - 1 )) p; $line_number_before q" $package_file
-	else
-		sed -n -i "1,$(( existing_line_number - 1 )) p; $existing_line_number q" $package_file
+		if [[ -z "$line" ]]; then
+			sed -n -i "1,$(( line_number_before - 1 )) p; $line_number_before q" $package_file
+		else
+			sed -n -i "1,$(( existing_line_number - 1 )) p; $existing_line_number q" $package_file
+		fi
 	fi
-fi
 
-# Append entries for $fdp_release to $package_file
-cat $new_package_list_file >> $package_file
+	# Append entries for $fdp_release to $package_file
+	cat $new_package_list_file >> $package_file
+fi
 
 # copy updated fdp_package_list.sh file to infra01
 #rpm -q nfs-utils || sudo dnf -y install nfs-utils
